Still getting false positives. I added more fake data to reflect this and recreate the bug, but no dice. Then I added the .uniq and the bug is gone. Unfortunately I can't just pop the data I found the bug with into git, because I'm using a few crufty client/ex-client apps to develop this on.
I'm going to have to find where this bug came from, I think, because there's a good chance the .uniq won't work when I expand the diff stuff. Right now the diff is interesting but you have to do it directly via the codebase model, you can't do it via a Towelie object. No nice views, either, and the lack of useful metadata is a problem. For the diff to be useful, at a minimum, it has to say what a given method definition is different *from*.